backend/requirements.txt
fastapi
pydantic
sqlalchemy
psycopg2-binary
bcrypt
uvicorn
pyjwt
python-multipart
passlib
stripe
python-dotenv
transformers
tensorflow
nltk
textblob
scikit-learn
pandas
numpy
